refactor shared metadata merging code into its own file
