+++ pipeline/metadata/beam_metadata.py
@@ -1,0 +1,57 @@
+"""Helpers for merging metadata with rows in beam pipelines"""
+
+from __future__ import absolute_import
+
+from typing import Tuple, Dict, List, Iterator
+
+from pipeline.metadata.flatten import Row
+
+# A key containing a date and IP
+# ex: ("2020-01-01", '1.2.3.4')
+DateIpKey = Tuple[str, str]
+
+# PCollection key names used internally by the beam pipeline
+IP_METADATA_PCOLLECTION_NAME = 'metadata'
+ROWS_PCOLLECION_NAME = 'rows'
+
+
+def merge_metadata_with_rows(  # pylint: disable=unused-argument
+    key: DateIpKey,
+    value: Dict[str, List[Row]],
+    field: str = None) -> Iterator[Row]:
+  # pyformat: disable
+  """Merge a list of rows with their corresponding metadata information.
+
+  Args:
+    key: The DateIpKey tuple that we joined on. This is thrown away.
+    value: A two-element dict
+      {IP_METADATA_PCOLLECTION_NAME: One element list containing an ipmetadata
+               ROWS_PCOLLECION_NAME: Many element list containing row dicts}
+      where ipmetadata is a dict of the format {column_name, value}
+       {'netblock': '1.0.0.1/24', 'asn': 13335, 'as_name': 'CLOUDFLARENET', ...}
+      and row is a dict of the format {column_name, value}
+       {'domain': 'test.com', 'ip': '1.1.1.1', 'success': true ...}
+    field: indicates a row field to update with metadata instead of the row (default).
+
+  Yields:
+    row dict {column_name, value} containing both row and metadata cols/values
+  """
+  # pyformat: enable
+  if value[IP_METADATA_PCOLLECTION_NAME]:
+    ip_metadata = value[IP_METADATA_PCOLLECTION_NAME][0]
+  else:
+    ip_metadata = {}
+  rows = value[ROWS_PCOLLECION_NAME]
+
+  for row in rows:
+    new_row: Row = {}
+    new_row.update(row)
+    if field == 'received':
+      if new_row['received']:
+        new_row['received'].update(ip_metadata)
+        new_row['received'].pop('date', None)
+        new_row['received'].pop('name', None)
+        new_row['received'].pop('country', None)
+    else:
+      new_row.update(ip_metadata)
+    yield new_row
